CalendarFragment.kt --> GalleryFragment.kt로 마이그레이션 - 캘린더 구현 준비
